+++ ChangeLog.md
@@ -3,7 +3,8 @@
 
 # NOT RELEASED
 
-
+- NEW : Display qty of current document  - *12/09/2024* - 1.8.0
+- NEW : Setup conf for default search behavior  - *12/09/2024* - 1.7.0
 
 ## Release 1.6
 - FIX : Permettre de lister les produits hors achat / hors vente - *25/09/2024* - 1.6.2